requirements.txt
pydantic>=2.5.0,
pdfplumber>=0.10.0,
pypdf>=4.0.0,
pdf2image>=1.16.0,
pillow>=10.0.0,
openai>=1.0.0,          # For OpenRouter compatibility
requests>=2.31.0,
langgraph>=0.1.0,
langchain>=0.1.0,
chromadb>=0.4.0,
faiss-cpu>=1.7.4,
sqlalchemy>=2.0.0,
pyyaml>=6.0,
numpy>=1.24.0,